cleaned up kick
