+++ Railgun/Commands/Inactivity/InactivityThresholdKick.cs
@@ -1,0 +1,49 @@
+﻿using Discord;
+using Finite.Commands;
+using Railgun.Core;
+using Railgun.Core.Attributes;
+using System.Threading.Tasks;
+using TreeDiagram;
+
+namespace Railgun.Commands.Inactivity
+{
+    public partial class Inactivity
+    {
+        public partial class Threshold
+        {
+            [Alias("kick"), BotPerms(GuildPermission.KickMembers)]
+            public class InactivityThresholdKick : SystemBase
+            {
+                [Command]
+                public Task ExecuteAsync(int threshold)
+                {
+                    if (threshold < 0)
+                        return ReplyAsync($"Inactivity Threshold can {Format.Bold("NOT")} be below 1!");
+
+                    var profile = Context.Database.ServerProfiles.GetOrCreateData(Context.Guild.Id);
+                    var data = profile.Inactivity;
+
+                    if (threshold == 0)
+                    {
+                        if (data.KickDaysThreshold == 0) 
+                            return ReplyAsync("Auto-Kick for inactivity is already turned off.");
+
+                        data.KickDaysThreshold = 0;
+
+                        return ReplyAsync("Auto-Kick for inactivity is now turned off.");
+                    }
+
+                    if (data.InactiveDaysThreshold < 1) 
+                        return ReplyAsync("Please set the Inactivity Threshold!");
+
+                    if (data.InactiveDaysThreshold > threshold) 
+                        return ReplyAsync($"Kick Threshold must be set higher than Inactivity Threshold ({Format.Bold(data.InactiveDaysThreshold.ToString())} day(s))!");
+
+                    data.KickDaysThreshold = threshold;
+
+                    return ReplyAsync($"Auto-Kick for inactivity is now turned on! Kick Threshold has been set to {Format.Bold(threshold.ToString())} days.");
+                }
+            }
+        }
+    }
+}
